[Refactor] useHighlights.tsx 모듈화
